moved all xml dataclasses to xml_utils to reduce length of read_xml
Also has the advantage of focusing read_xml down to just the FreeCAD XML classes instead of mixing them with the dataclasses.
